+++ Source/glTFForUE4Ed/Private/glTF/glTFImporterEdStaticMesh.cpp
@@ -110,7 +110,7 @@ UStaticMesh* FglTFImporterEdStaticMesh::CreateStaticMesh(const TWeakPtr<FglTFImp
                 continue;
             }
             const int32 NodeId = *NodeIdPtr;
-            if (NodeId < 0 || NodeId >= InGlTF->nodes.size())
+            if (NodeId < 0 || NodeId >= static_cast<int32>(InGlTF->nodes.size()))
             {
                 checkSlow(0);
                 continue;
@@ -217,7 +217,7 @@ bool FglTFImporterEdStaticMesh::GenerateRawMesh(const std::shared_ptr<libgltf::S
     if (!!(InNode->mesh))
     {
         const int32_t MeshId = *(InNode->mesh);
-        if (MeshId < 0 || MeshId >= InGlTF->meshes.size()) return false;
+        if (MeshId < 0 || MeshId >= static_cast<int32>(InGlTF->meshes.size())) return false;
         const auto& Mesh = InGlTF->meshes[MeshId];
         if (!GenerateRawMesh(InGlTF, Mesh, InNodeAbsoluteTransform, InBuffers, OutRawMesh, InOutglTFMaterialInfos, InFeedbackTaskWrapper))
         {
@@ -234,7 +234,7 @@ bool FglTFImporterEdStaticMesh::GenerateRawMesh(const std::shared_ptr<libgltf::S
             continue;
         }
         const int32 NodeId = *NodeIdPtr;
-        if (NodeId < 0 || NodeId >= InGlTF->nodes.size())
+        if (NodeId < 0 || NodeId >= static_cast<int32>(InGlTF->nodes.size()))
         {
             checkSlow(0);
             continue;
@@ -266,7 +266,7 @@ bool FglTFImporterEdStaticMesh::GenerateRawMesh(const std::shared_ptr<libgltf::S
     if (!InMesh) return false;
 
     FString MeshName = InMesh->name.c_str();
-    for (int32 i = 0; i < InMesh->primitives.size(); ++i)
+    for (int32 i = 0; i < static_cast<int32>(InMesh->primitives.size()); ++i)
     {
         const auto& Primitive = InMesh->primitives[i];
         FRawMesh NewRawMesh;
